opengametools: initial code for ogt_vox_loader
ogt_vox_loader is a lightweight, low-dependency
library to get at the scene data within a .vox file
that was written by MagicaVoxel.

The library allows callers to pass a .vox
file in a memory buffer to create a scene
structure: ogt_vox_scene.

The ogt_vox_scene can be parsed to read:
* models:
  - a conceptual grid of voxel data with palette indices
* palette:
  - defines colors for each color index
* instances:
  - name
  - transforms within the scene
  - model index
  - hidden/shown state.
  - layer index
* layers:
  - layer name
  - hidden/shown state