ogt: ogt_vox_loader now has ability to save out to a .vox file.
I currently have too many .vox files on disk, each containing a single voxel model, which is laborious to edit. To solve this, I'm planning to build a tool to merge them all into a single .vox file and do a smart-merge on the palettes. After they are merged, I'd need the ability to save the merged scene back out to disk.  This commit therefore adds the ability to save an ogt_vox_scene back out to a .vox file that is loadable by MagicaVoxel 0.99.3-alpha

- Callers can now use ogt_vox_write_scene to generate .vox file into a memory buffer that can then be saved out to disk.
- Note: this won't produce binary identical .vox files.
- Renamed ogt_vox_create_scene -> ogt_vox_read_scene to more obviously mirror the new ogt_vox_write_scene api.
- Renamed some vars using old style casing. whoops.
- Fixed up some tabs -> spaces
- ogt_vox_loader.h will be renamed in a subsequent commit.