ogt: added ogt_vox_merge_scenes API
Added new ogt_vox_merge_scenes which takes an array of ogt_vox_scene pointers and constructs a new scene that contains all of them. If the client then uses ogt_vox_write_scene on the merged scene, this allows the client to consolidate multiple .vox files into a single one that is loadable by Magicavoxel.

The important step taken here is this merge function also takes care of palette remapping. The merged scene palette will contain only the parts of the source scene palettes that are actually used by the models within each scene. If by doing this we run out of entries in the merged palette, we start remapping to existing colors based on finding the closest euclidean distance in RGB space between the used color and the merged palette color.